wifi: Do not set wlan.driver.status to ok after insmod for LONU
The check for wlan.driver.status in is_wifi_driver_loaded does 
expects this property to be not "OK" for LONU to trigger 
wifi_change_driver_state. This property is later set to "OK" 
from wifi_load_driver. Thus , do not set this property to "OK" 
after the insmod during the boot.

Change-Id: I4d54149de9c9bd19cd832146c5af903f45dde340
CRs-Fixed: 2264381